# Improve import performance